[Feature] Readout noise (#286)
Adding noise readout (issue #285). I am thinking initially Readout would
be specified at the level of the circuit. I am though unsure if this is
a good idea given the current Qadence interface of Readout. I wonder if
it makes sense to add it as a noise protocol but it is not compatible
with the other noise signatures we have. What do you think @jpmoutinho
@RolandMacDoland ?

I also fixed the equations display in `noise.md` and remove the annoying
logging with Hamevo (it was printing too much on qadence's side).

---------